Refactor player controller into namespace
