+++ .gitignore
@@ -22,6 +22,9 @@
 # Visual Studio cache directory
 .vs/
 
+# Rider files
+.idea/
+
 # Gradle cache directory
 .gradle/
