+++ ProjectSettings/Packages/com.unity.settings-manager/Settings.json
@@ -1,0 +1,7 @@
+{
+    "m_Name": "Settings",
+    "m_Path": "ProjectSettings/Packages/com.unity.settings-manager/Settings.json",
+    "m_Dictionary": {
+        "m_DictionaryValues": []
+    }
+}
